// src/definition/korean-open-api.ts
export enum LanguageCode {
  english = 1,
  japanese = 2,
  french = 3,
  spanish = 4,
  arabic = 5,
  mongolian = 6,
  vietnamese = 7,
  thai = 8,
  indonesian = 9,
  russian = 10,
}

export enum WordGrade {
  none = '없음',
  beginner = '초급',
  intermediate = '중급',
  advanced = '고급'
}

export enum LanguageNames {
  english = "영어",
  japanese = "일본어",
  french = "프랑스어",
  spanish = "스페인어",
  arabic = "아랍어",
  mongolian = "몽골어",
  vietnamese = "베트남어",
  thai = "타이어",
  indonesian = "인도네시아어",
  russian = "러시아어",
}